Printed out the minutes as a decimal number

+++ main/time/debug.c
@@ -2,9 +2,9 @@
 
 void printEventMinutes(int eventNum, struct timeval time)
 {
-  int64_t timeMicro = timevalToMicro(time);
-  int64_t timeMinutes = US_TO_MINUTES(timeMicro);
-  otLogNotePlat("The %dth event will be in %" PRId64 " minutes.",
+  double timeMicro = (double) timevalToMicro(time);
+  double timeMinutes = US_TO_MINUTES(timeMicro);
+  otLogNotePlat("The %dth event will be in %.3f minutes.",
                 eventNum, timeMinutes);
   return;
 }
